@@ -1,26 +1,30 @@
 {
   "name": "functions",
-  "description": "Cloud Functions for Firebase",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
-    "serve": "firebase serve --only functions",
-    "shell": "firebase functions:shell",
+    "build": "tsc",
+    "serve": "npm run build && firebase emulators:start --only functions",
+    "shell": "npm run build && firebase functions:shell",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "12"
+    "node": "14"
   },
+  "main": "lib/index.js",
   "dependencies": {
-    "date-fns": "2.6.0",
+    "date-fns": "2.23.0",
     "deep-diff": "1.0.2",
-    "firebase-admin": "8.7.0",
-    "firebase-functions": "3.3.0",
-    "nodemailer": "6.3.1",
-    "nodemailer-smtp-transport": "2.7.4"
+    "firebase-admin": "9.11.1",
+    "firebase-functions": "3.15.4",
+    "nodemailer": "6.6.3"
   },
   "devDependencies": {
-    "firebase-functions-test": "0.1.7"
+    "@types/deep-diff": "1.0.1",
+    "@types/nodemailer": "6.4.4",
+    "firebase-functions-test": "0.3.2",
+    "typescript": "4.4.2"
   }
 }
